--- src/main/java/AppOptionException.java
@@ -1,10 +1,0 @@
-import java.io.Serial;
-
-abstract class AppOptionException extends AppException {
-    @Serial
-    private static final long serialVersionUID = 0L;
-
-    AppOptionException(final String message) {
-        super(message);
-    }
-}
