--- src/main/java/AppStateException.java
@@ -1,10 +1,0 @@
-import java.io.Serial;
-
-abstract class AppStateException extends AppException {
-    @Serial
-    private static final long serialVersionUID = 0L;
-
-    AppStateException(final String message) {
-        super(message);
-    }
-}
