--- src/main/java/Event.java
@@ -1,16 +1,0 @@
-final class Event extends Task {
-    private final String from;
-    private final String to;
-
-    Event(final String name, final String from, final String to) {
-        super(name);
-
-        this.from = from;
-        this.to = to;
-    }
-
-    @Override
-    public String toString() {
-        return "[E]" + super.toString() + " (from: " + from + ", to: " + to + ")";
-    }
-}
