--- src/main/java/Todo.java
@@ -1,10 +1,0 @@
-final class Todo extends Task {
-    Todo(final String name) {
-        super(name);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
