+++ src/main/java/arpa/home/yikjin/app/kuro/command/InputCommand.java
@@ -1,0 +1,27 @@
+package arpa.home.yikjin.app.kuro.command;
+
+enum InputCommand {
+    BYE("bye"), DEADLINE("deadline"), EVENT("event"), LIST("list"), MARK("mark"), TODO("todo"), UNMARK(
+            "unmark"), INVALID("");
+
+    private final String enumValue;
+
+    InputCommand(final String enumValue) {
+        this.enumValue = enumValue;
+    }
+
+    static InputCommand parse(final String command) {
+        for (final InputCommand enumCommand : values()) {
+            if (enumCommand.enumValue.equalsIgnoreCase(command)) {
+                return enumCommand;
+            }
+        }
+
+        return INVALID;
+    }
+
+    @Override
+    public String toString() {
+        return enumValue;
+    }
+}
