+++ src/main/java/arpa/home/yikjin/app/kuro/exception/base/AppOptionException.java
@@ -1,0 +1,12 @@
+package arpa.home.yikjin.app.kuro.exception.base;
+
+import java.io.Serial;
+
+public abstract class AppOptionException extends AppException {
+    @Serial
+    private static final long serialVersionUID = 0L;
+
+    protected AppOptionException(final String message) {
+        super(message);
+    }
+}
