+++ src/main/java/arpa/home/yikjin/app/kuro/exception/input/argument/package-info.java
@@ -1,0 +1,1 @@
+package arpa.home.yikjin.app.kuro.exception.input.argument;
